Fix python & clippy warnings
