+++ libafl_qemu/src/helpers/edges.rs
@@ -539,8 +539,7 @@ where
             let paging_id = hooks
                 .qemu()
                 .current_cpu()
-                .map(|cpu| cpu.current_paging_id())
-                .flatten();
+                .and_then(|cpu| cpu.current_paging_id());
 
             if !h.must_instrument(src, paging_id) && !h.must_instrument(dest, paging_id) {
                 return None;
@@ -609,8 +608,7 @@ where
             let paging_id = hooks
                 .qemu()
                 .current_cpu()
-                .map(|cpu| cpu.current_paging_id())
-                .flatten();
+                .and_then(|cpu| cpu.current_paging_id());
 
             if !h.must_instrument(src, paging_id) && !h.must_instrument(dest, paging_id) {
                 return None;
@@ -676,8 +674,7 @@ where
             let paging_id = hooks
                 .qemu()
                 .current_cpu()
-                .map(|cpu| cpu.current_paging_id())
-                .flatten();
+                .and_then(|cpu| cpu.current_paging_id());
 
             if !h.must_instrument(pc, paging_id) {
                 return None;
