+++ libafl_qemu/src/qemu/mod.rs
@@ -532,8 +532,8 @@ impl Qemu {
             {
                 qemu_init(
                     argc,
-                    argv.as_ptr() as *const *const u8,
-                    envp.as_ptr() as *const *const u8,
+                    argv.as_ptr(),
+                    envp.as_ptr(),
                 );
                 libc::atexit(qemu_cleanup_atexit);
                 libafl_qemu_sys::syx_snapshot_init(true);
@@ -977,49 +977,11 @@ impl ArchExtras for Qemu {
 
 #[cfg(feature = "python")]
 pub mod pybind {
-    use pyo3::{exceptions::PyValueError, prelude::*, types::PyInt};
+    use pyo3::{exceptions::PyValueError, prelude::*};
 
-    use super::{GuestAddr, GuestUsize, MmapPerms, SyscallHookResult};
+    use super::{GuestAddr, GuestUsize};
 
-    static mut PY_SYSCALL_HOOK: Option<PyObject> = None;
     static mut PY_GENERIC_HOOKS: Vec<(GuestAddr, PyObject)> = vec![];
-
-    extern "C" fn py_syscall_hook_wrapper(
-        _data: u64,
-        sys_num: i32,
-        a0: u64,
-        a1: u64,
-        a2: u64,
-        a3: u64,
-        a4: u64,
-        a5: u64,
-        a6: u64,
-        a7: u64,
-    ) -> SyscallHookResult {
-        unsafe { PY_SYSCALL_HOOK.as_ref() }.map_or_else(
-            || SyscallHookResult::new(None),
-            |obj| {
-                let args = (sys_num, a0, a1, a2, a3, a4, a5, a6, a7);
-                Python::with_gil(|py| {
-                    let ret = obj.call1(py, args).expect("Error in the syscall hook");
-                    let any = ret.as_ref(py);
-                    if any.is_none() {
-                        SyscallHookResult::new(None)
-                    } else {
-                        let a: Result<&PyInt, _> = any.downcast();
-                        if let Ok(i) = a {
-                            SyscallHookResult::new(Some(
-                                i.extract().expect("Invalid syscall hook return value"),
-                            ))
-                        } else {
-                            SyscallHookResult::extract(any)
-                                .expect("The syscall hook must return a SyscallHookResult")
-                        }
-                    }
-                })
-            },
-        )
-    }
 
     extern "C" fn py_generic_hook_wrapper(idx: u64, _pc: GuestAddr) {
         let obj = unsafe { &PY_GENERIC_HOOKS[idx as usize].1 };
@@ -1088,66 +1050,8 @@ pub mod pybind {
             self.qemu.remove_breakpoint(addr);
         }
 
-        fn g2h(&self, addr: GuestAddr) -> u64 {
-            self.qemu.g2h::<*const u8>(addr) as u64
-        }
-
-        fn h2g(&self, addr: u64) -> GuestAddr {
-            self.qemu.h2g(addr as *const u8)
-        }
-
-        fn binary_path(&self) -> String {
-            self.qemu.binary_path().to_owned()
-        }
-
-        fn load_addr(&self) -> GuestAddr {
-            self.qemu.load_addr()
-        }
-
         fn flush_jit(&self) {
             self.qemu.flush_jit();
-        }
-
-        fn map_private(&self, addr: GuestAddr, size: usize, perms: i32) -> PyResult<GuestAddr> {
-            if let Ok(p) = MmapPerms::try_from(perms) {
-                self.qemu
-                    .map_private(addr, size, p)
-                    .map_err(PyValueError::new_err)
-            } else {
-                Err(PyValueError::new_err("Invalid perms"))
-            }
-        }
-
-        fn map_fixed(&self, addr: GuestAddr, size: usize, perms: i32) -> PyResult<GuestAddr> {
-            if let Ok(p) = MmapPerms::try_from(perms) {
-                self.qemu
-                    .map_fixed(addr, size, p)
-                    .map_err(PyValueError::new_err)
-            } else {
-                Err(PyValueError::new_err("Invalid perms"))
-            }
-        }
-
-        fn mprotect(&self, addr: GuestAddr, size: usize, perms: i32) -> PyResult<()> {
-            if let Ok(p) = MmapPerms::try_from(perms) {
-                self.qemu
-                    .mprotect(addr, size, p)
-                    .map_err(PyValueError::new_err)
-            } else {
-                Err(PyValueError::new_err("Invalid perms"))
-            }
-        }
-
-        fn unmap(&self, addr: GuestAddr, size: usize) -> PyResult<()> {
-            self.qemu.unmap(addr, size).map_err(PyValueError::new_err)
-        }
-
-        fn set_syscall_hook(&self, hook: PyObject) {
-            unsafe {
-                PY_SYSCALL_HOOK = Some(hook);
-            }
-            self.qemu
-                .add_pre_syscall_hook(0u64, py_syscall_hook_wrapper);
         }
 
         fn set_hook(&self, addr: GuestAddr, hook: PyObject) {
